+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Default parameter set for `Connect-Instance`.
+
 ## [1.0.0] - 2022-08-18
 
 ### Added